+++ examples/08_advanced_ml_scanners.py
@@ -8,9 +8,7 @@ This example shows how to configure DocFirewall to:
 """
 
 import os
-import sys
 
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../src")))
 from doc_firewall import ScanConfig, Scanner